requirements.txt
tornado==4.3
tornadoes==2.4.2
elasticsearch==2.3.0
python-dateutil==2.5.3
Pillow==3.2.0
numpy==1.11.0
ImageHash==1.0
requests==2.10.0
scikit-learn==0.17
hyperopt==0.0.2
#required by hyperopt:
pymongo==3.2.2
base58==0.2.2
mediachain-client>=0.1.14
asteval==0.9.7
ujson==1.33
editdistance==0.3.1
lmdb==0.89
annoy==1.8.0
Theano==0.9.0.dev1
scikit-image==0.12.3
scipy=0.17.1
caffe=1.0.0-rc3
scandir==1.3
tensorflow==0.10.0rc0
